backend/requirements/local.txt
-r base.txt

pytest-django==4.5.2

requests==2.28.2
